Separate trap handler from common.c
