+++ sw/common/common.c
@@ -1,7 +1,5 @@
 #include <stdint.h>
-#include "FreeRTOS.h"
-#include "task.h"
-#include "timers.h"
+
 #include "uart.h"
 #include "printf.h"
 #include "common.h"
@@ -15,26 +13,6 @@ void _putchar(char character)
         uart_putchar('\r');
     }
     uart_putchar(character);
-}
-
-/*
- * Application exception handler for FreeRTOS
- */
-void freertos_risc_v_application_exception_handler( uint32_t cause )
-{
-    printf("Application Exception Handler Called By FreeRTOS!!!\n");
-    printf("MCAUSE: 0x%x\n", cause);
-    return;
-}
-
-/*
- * Application interrupt handler for FreeRTOS
- */
-void freertos_risc_v_application_interrupt_handler( uint32_t cause )
-{
-    printf("Application Interrupt Handler Called By FreeRTOS!!!\n");
-    printf("MCAUSE: 0x%x\n", cause);
-    return;
 }
 
 /*
